requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0
sentence-transformers==2.2.2
chromadb==0.4.18
mysql-connector-python==8.2.0
pandas==2.1.3
numpy==1.26.2
rapidfuzz==3.5.2
python-dotenv==1.0.0